+++ src/Strategies/Scanner.js
@@ -2,6 +2,7 @@ import OrderController from "../Controllers/OrderController.js";
 import AccountStore from "../Store/AccountStore.js";
 import Markets from "../Backpack/Markets.js";
 import Utils from "../Utils/Utils.js";
+import { logInfo } from "../Utils/logger.js";
 
 class Scanner {
   constructor() {
@@ -21,80 +22,83 @@ class Scanner {
     this.newMarkets = newMarketsBKP ? newMarketsBKP : [];
   }
 
-  async _doMarketTrade(newMarket, side = "random") {
-    const markPrices = await Markets.getAllMarkPrices(newMarket.symbol);
-    const marketPrice = parseFloat(markPrices[0].markPrice);
-    newMarket.price = marketPrice;
+  async _doPerpTrade(newMarket, side = "random") {
     let order = {};
     order.symbol = newMarket.symbol;
-    order.entry = marketPrice;
+    order.entry = newMarket.price;
     order.decimal_quantity = newMarket.decimal_quantity;
     order.decimal_price = newMarket.decimal_price;
     order.stepSize_quantity = newMarket.stepSize_quantity;
     order.tickSize = newMarket.tickSize;
-    order.volume = 10;
+    order.quantity = newMarket.quantity;
 
     if (side === "long") {
-      order.stop = marketPrice * 0.99;
-      order.target = marketPrice * 1.01;
+      order.stop = newMarket.price * 0.99;
+      order.target = newMarket.price * 1.01;
       order.action = "long";
     } //
     else if (side === "short") {
-      order.stop = marketPrice * 1.01;
-      order.target = marketPrice * 0.99;
+      order.stop = newMarket.price * 1.01;
+      order.target = newMarket.price * 0.99;
       order.action = "short";
     } //
     else {
       const randomBoolean = Math.random() >= 0.5;
       if (randomBoolean) {
-        order.stop = marketPrice * 0.99;
-        order.target = marketPrice * 1.01;
+        order.stop = newMarket.price * 0.99;
+        order.target = newMarket.price * 1.01;
         order.action = "long";
       } else {
-        order.stop = marketPrice * 1.01;
-        order.target = marketPrice * 0.99;
+        order.stop = newMarket.price * 1.01;
+        order.target = newMarket.price * 0.99;
         order.action = "short";
       }
     }
-    await OrderController.openMarketOrderScanner(order);
+    return await OrderController.openMarketOrder(order);
   }
 
-  _generateOrdersBatch(newMarket, number) {
+  async _doSpotTrade(side, newMarket, volume) {
+    const order = {};
+    order.side = side;
+    order.volume = volume;
+    order.symbol = newMarket.symbol;
+    order.decimal_quantity = newMarket.decimal_quantity;
+    order.decimal_price = newMarket.decimal_price;
+    order.stepSize_quantity = newMarket.stepSize_quantity;
+    order.tickSize = newMarket.tickSize;
+    return await OrderController.openOrderSpot(order);
+  }
+
+  async _doBatchTrade(newMarket, number) {
     const ordersBatch = [];
     for (let i = -number; i <= number; i++) {
       if (i === 0) continue;
       let order = {};
       order.symbol = newMarket.symbol;
-      // order.decimal_quantity = newMarket.decimal_quantity;
-      // order.decimal_price = newMarket.decimal_price;
-      // order.stepSize_quantity = newMarket.stepSize_quantity;
-      // order.tickSize = newMarket.tickSize;
-      order.decimal_quantity = newMarket.qtdHouses;
-      order.decimal_price = newMarket.prcHouses;
-      order.stepSize_quantity = newMarket.qtdStep;
-      order.tickSize = newMarket.prcStep;
-      order.volume = 1;
+      order.decimal_quantity = newMarket.decimal_quantity;
+      order.decimal_price = newMarket.decimal_price;
+      order.stepSize_quantity = newMarket.stepSize_quantity;
+      order.tickSize = newMarket.tickSize;
+      order.quantity = newMarket.quantity;
       const entryPrice = newMarket.price + parseFloat(newMarket.tickSize) * i;
       order.entry = entryPrice;
       if (i > 0) {
         order.action = "long";
-        order.stop = entryPrice - parseFloat(newMarket.tickSize) * 5;
-        order.target = entryPrice + parseFloat(newMarket.tickSize) * 5;
+        //order.stop = entryPrice - parseFloat(newMarket.tickSize) * 5;
+        //order.target = entryPrice + parseFloat(newMarket.tickSize) * 5;
       } else {
         order.action = "short";
-        order.stop = entryPrice + parseFloat(newMarket.tickSize) * 5;
-        order.target = entryPrice - parseFloat(newMarket.tickSize) * 5;
+        //order.stop = entryPrice + parseFloat(newMarket.tickSize) * 5;
+        //order.target = entryPrice - parseFloat(newMarket.tickSize) * 5;
       }
       ordersBatch.push(order);
     }
-    return ordersBatch;
+    return await OrderController.createBatchOfMarketTriggerOrders(ordersBatch);
   }
 
   async run() {
     try {
       if (!this.initialized) await this.init();
-
-      console.log("\n📣 Previous candle closed. Running a new cicle of analysis.\n");
 
       //Check sufficient account balance for new orders
       const capitalAvailable = await AccountStore.getAvailableCapital();
@@ -103,8 +107,9 @@ class Scanner {
         return;
       }
 
-      const currentMarkets = await AccountStore.getMarkets();
-      const newMarkets = currentMarkets.filter((el) => {
+      // Check PERP Markets
+      const currentPerpMarkets = await AccountStore.getMarkets();
+      const newPerpMarkets = currentPerpMarkets.filter((el) => {
         const isNew = !this.knownMarkets.includes(el.symbol);
         if (isNew) {
           this.knownMarkets.push(el.symbol);
@@ -113,8 +118,9 @@ class Scanner {
             totalTrades: 0,
             clockinNextTime: null,
             clockinOrdersSent: 0,
-            doLong: false,
             phase: "first50",
+            type: "perp",
+            quantity: null,
             decimal_quantity: el.decimal_quantity,
             decimal_price: el.decimal_price,
             stepSize_quantity: el.stepSize_quantity,
@@ -123,88 +129,232 @@ class Scanner {
         }
         return isNew;
       });
-
-      if (newMarkets.length > 0) {
-        console.log("🌟 New PREP market(s) FOUND!", newMarkets);
-        for (const market of newMarkets) {
-          await Utils.notify("🌟 New PERP token(s) found! " + market.symbol + "\nInitiating the routine.");
-        }
+      if (newPerpMarkets.length > 0) {
+        console.log("🌟 New PREP market(s) FOUND!", newPerpMarkets);
         await Utils.saveDataToFile(this.knownMarkets, "knownPerpMarkets.json");
-      } else {
-        console.log("No new PERP market found.");
+        for (const market of newPerpMarkets) {
+          await Utils.notify("🌟 New PERP token(s) found! " + market.symbol);
+        }
       }
 
+      // Check SPOT markets
+      const currentSpotMarkets = await AccountStore.getMarkets("SPOT");
+      const newSpotMarkets = currentSpotMarkets.filter((el) => {
+        const isNew = !this.knownSpotMarkets.includes(el.symbol);
+        if (isNew) {
+          this.knownSpotMarkets.push(el.symbol);
+          this.newMarkets.push({
+            symbol: el.symbol,
+            totalTrades: 0,
+            clockinNextTime: null,
+            clockinOrdersSent: 0,
+            phase: "first50",
+            type: "spot",
+            volume: null,
+            decimal_quantity: el.decimal_quantity,
+            decimal_price: el.decimal_price,
+            stepSize_quantity: el.stepSize_quantity,
+            tickSize: el.tickSize,
+          });
+        }
+        return isNew;
+      });
+      if (newSpotMarkets.length > 0) {
+        console.log("🌟 New SPOT market(s) FOUND!", newSpotMarkets);
+        await Utils.saveDataToFile(this.knownSpotMarkets, "knownSpotMarkets.json");
+        for (const spotMarket of newSpotMarkets) {
+          await Utils.notify("🌟 New SPOT token found! " + spotMarket.symbol);
+        }
+      }
+
+      //= = = = = = = = = = = = = = = =//
+      //                               //
+      // newly found markets routines  //
+      //                               //
+      //= = = = = = = = = = = = = = = =//
       if (this.newMarkets.length > 0) {
-        for (let i = this.newMarkets.length - 1; i >= 0; i--) {
-          const newMarket = this.newMarkets[i];
-          // First 50 routine: try to be one of 50 to trade the new token
+        //
+        // First 50 routine: try to be one of 50 to trade the new token
+        //
+        let first50Length = 0;
+        for (const newMarket of this.newMarkets) {
           if (newMarket.phase === "first50") {
             console.log("1️⃣  Executing 'first50' routine for " + newMarket.symbol);
-            await this._doMarketTrade(newMarket, "long");
-            newMarket.phase = "lucky777";
+
+            // const markPrices = await Markets.getAllMarkPrices(newMarket.symbol);
+            // const marketPrice = parseFloat(markPrices[0].markPrice);
+            // newMarket.price = marketPrice;
+
+            const candles = await Markets.getKLines(newMarket.symbol, "1m", 3);
+            let marketPrice = null;
+            if (candles) marketPrice = candles[2]?.close || candles[1]?.close || candles[0]?.close || null;
+            newMarket.price = parseFloat(marketPrice);
+            console.log(newMarket.symbol, newMarket.price);
+
+            //infer props
+            if (newMarket.decimal_quantity === undefined || newMarket.stepSize_quantity === 0) {
+              const props = Utils.inferMarketProps(marketPrice);
+              newMarket.decimal_quantity = props.qtdHouses;
+              newMarket.decimal_price = props.prcHouses;
+              newMarket.stepSize_quantity = props.qtdStep;
+              newMarket.tickSize = props.prcStep;
+            }
+
+            //estimated minimum quantity allowed
+            const x = 1 / marketPrice;
+            const n = Math.floor(Math.log10(x));
+            newMarket.quantity = Math.pow(10, n);
+
+            newMarket.attemptsLeft = 3;
+            first50Length += 1;
           }
+        }
+        if (first50Length > 0) {
+          let finished = 0;
+          const spotVolumes = ["15", "7", "3"];
+
+          for (let i = 0; i < 100; i++) {
+            let j = i % first50Length;
+            const newMarket = this.newMarkets[j];
+
+            if (newMarket.attemptsLeft > 0 && newMarket.phase === "first50") {
+              if (newMarket.type === "perp") {
+                newMarket.attemptsLeft -= 1;
+                const response = await this._doPerpTrade(newMarket, "long");
+                if (typeof response === "object" && response.status === "Filled") {
+                  Utils.notify(`✅ Successfully traded ${newMarket.symbol}`);
+                  newMarket.phase = "lucky777";
+                  // close successful perp trade
+                  setTimeout(() => {
+                    this._doPerpTrade(newMarket, "short");
+                  }, 5000);
+                  finished += 1;
+                } else if (
+                  newMarket.attemptsLeft > 0 &&
+                  (response?.message === "Quantity is below the minimum allowed value" ||
+                    response?.message === "Quantity decimal too long")
+                ) {
+                  //'Market orders must specify a positive `quantity`' //executedQuantity: '100'
+                  console.log(`${newMarket.symbol} trade failed. Trying again with more quantity...`);
+                  newMarket.quantity *= 10;
+                } else {
+                  const msg = `❌ Failed to trade ${newMarket.symbol}. Reason: ${response?.message}`;
+                  Utils.notify(msg);
+                  console.log(msg);
+                  logInfo(msg);
+                  newMarket.phase = "invalid";
+                  finished += 1;
+                }
+              } //
+              else if (newMarket.type === "spot") {
+                newMarket.attemptsLeft -= 1;
+                const volume = spotVolumes[newMarket.attemptsLeft];
+                const response = await this._doSpotTrade("Bid", newMarket, volume);
+                if (typeof response === "object" && response.status === "Filled") {
+                  Utils.notify(`✅ Successfully bought $${volume} of ${newMarket.symbol}`);
+                  newMarket.volume = volume;
+                  newMarket.phase = "clockingin";
+                  newMarket.clockinNextTime = Date.now() + 120000;
+                  console.log(newMarket.symbol + " set to 'clockingin' phase.");
+                  // close successful spot trade
+                  setTimeout(() => {
+                    this._doSpotTrade("Ask", newMarket, response.executedQuoteQuantity * 0.999);
+                  }, 5000);
+                  finished += 1;
+                } //
+                else if (
+                  newMarket.attemptsLeft > 0 &&
+                  response?.message === "Quantity is below the minimum allowed value"
+                ) {
+                  console.log("Trade failed. Trying again with more volume...");
+                } else {
+                  const msg = `❌ Failed to trade ${newMarket.symbol}. Reason: ${response?.message}`;
+                  Utils.notify(msg);
+                  console.log(msg);
+                  logInfo(msg);
+                  newMarket.phase = "invalid";
+                  finished += 1;
+                }
+              }
+
+              if (finished >= first50Length) {
+                break;
+              }
+            }
+          }
+          //remove the invalid ones from the list
+          for (let i = this.newMarkets.length - 1; i >= 0; i--) {
+            if (this.newMarkets[i].phase === "invalid") {
+              this.newMarkets.splice(i, 1);
+            }
+          }
+        }
+
+        for (let i = this.newMarkets.length - 1; i >= 0; i--) {
+          const newMarket = this.newMarkets[i];
+          //
           // Lucky 777 routine: try to do the 777 trade on the new token
+          //
+          //
           if (newMarket.phase === "lucky777") {
             console.log("2️⃣  Executing 'lucky777' routine for " + newMarket.symbol);
-            if (!newMarket.price) {
+            if (newMarket.type === "perp") {
+              await new Promise((resolve) => setTimeout(resolve, 2000));
               await OrderController.cancelAllOrders(newMarket.symbol);
-              const candles = await Markets.getKLines(newMarket.symbol, "1m", 1);
+
               const markPrices = await Markets.getAllMarkPrices(newMarket.symbol);
               newMarket.price = parseFloat(markPrices[0].markPrice);
-              if (!newMarket.doLong) {
-                await this._doMarketTrade(newMarket, "short");
-                newMarket.doLong = true;
+              const response = await this._doPerpTrade(newMarket, "long");
+              if (typeof response === "object" && response.status === "Filled") {
+                setTimeout(() => {
+                  this._doPerpTrade(newMarket, "short");
+                }, 5000);
               }
-              const trades = Number(candles[0].trades);
+              await this._doBatchTrade(newMarket, 5);
+
+              const candles = await Markets.getKLines(newMarket.symbol, "1m", 1);
+              const trades = candles ? Number(candles[0]?.trades) : null;
               if (trades) newMarket.totalTrades += trades;
-              console.log("totalTrades: ", newMarket.totalTrades);
+              console.log(`${newMarket.symbol} total trades: `, newMarket.totalTrades);
+
               if (newMarket.totalTrades > 777) {
                 newMarket.phase = "clockingin";
-                newMarket.clockinOrdersSent += 1;
-                newMarket.clockinNextTime = Date.now() + 43200000;
+                newMarket.clockinNextTime = Date.now() + 120000;
                 console.log(newMarket.symbol + " set to 'clockingin' phase.");
               }
-            } else {
-              const qtt = 1 / newMarket.price;
-              newMarket.qtdHouses = 5;
-              for (let i = 0; i < 5; i++) {
-                if ((qtt * 10 ** i).toFixed() >= 1) {
-                  newMarket.qtdHouses = i;
-                  break;
-                }
-              }
-              newMarket.prcHouses = 6;
-              for (let i = 1; i < 6; i++) {
-                if ((newMarket.price * 10 ** i).toFixed() >= 1000) {
-                  newMarket.prcHouses = i;
-                  break;
-                }
-              }
-              newMarket.prcStep = 1 / 10 ** newMarket.prcHouses;
-              newMarket.qtdStep = 1 / 10 ** newMarket.qtdHouses;
             }
-            const ordersBatch = this._generateOrdersBatch(newMarket, 5);
-            await OrderController.createBatchOfMarketTriggerOrders(ordersBatch);
-            newMarket.price = null;
           }
+
+          //
           // Clocking In: trade the new token for all 7 days after launch
+          //
           if (newMarket.phase === "clockingin") {
-            console.log(
-              `Clocking In: ${newMarket.symbol}   trades done: [${
-                newMarket.clockinOrdersSent
-              }/15]   next: ${Utils.formatDateTime(newMarket.clockinNextTime)}`
-            );
             if (Date.now() > newMarket.clockinNextTime) {
               console.log("3️⃣  Executing 'clocking in' routine for " + newMarket.symbol);
-              if (newMarket.doLong) {
-                await this._doMarketTrade(newMarket, "long");
-                newMarket.doLong = !newMarket.doLong;
-              } else {
-                await this._doMarketTrade(newMarket, "short");
-                newMarket.doLong = !newMarket.doLong;
+
+              if (newMarket.type === "perp") {
+                const response = await this._doPerpTrade(newMarket, "long");
+                if (typeof response === "object" && response.status === "Filled") {
+                  setTimeout(() => {
+                    this._doPerpTrade(newMarket, "short");
+                  }, 5000);
+                }
+              } //
+              else if (newMarket.type === "spot") {
+                const response = await this._doSpotTrade("Bid", newMarket, newMarket.volume);
+                if (typeof response === "object" && response.status === "Filled") {
+                  setTimeout(() => {
+                    this._doSpotTrade("Ask", newMarket, response.executedQuoteQuantity * 0.999);
+                  }, 5000);
+                }
               }
               newMarket.clockinOrdersSent += 1;
               newMarket.clockinNextTime += 43200000;
+              console.log(
+                `Clocking In: ${newMarket.symbol}   trades done: [${
+                  newMarket.clockinOrdersSent
+                }/15]   next: ${Utils.formatDateTime(newMarket.clockinNextTime)}`
+              );
               if (newMarket.clockinOrdersSent >= 15) {
                 this.newMarkets.splice(i, 1);
                 console.log("❇️  Completed all routines for " + newMarket.symbol);
@@ -219,37 +369,8 @@ class Scanner {
         }
       }
 
-      // Check SPOT markets
-      const currentSpotMarkets = await AccountStore.getMarkets("SPOT");
-      const newSpotMarkets = currentSpotMarkets.filter((el) => {
-        const isNew = !this.knownSpotMarkets.includes(el.symbol);
-        if (isNew) {
-          this.knownSpotMarkets.push(el.symbol);
-        }
-        return isNew;
-      });
-
-      if (newSpotMarkets.length > 0) {
-        console.log("🌟 New SPOT market(s) FOUND!", newSpotMarkets);
-        await Utils.saveDataToFile(this.knownSpotMarkets, "knownSpotMarkets.json");
-
-        for (const spotMarket of newSpotMarkets) {
-          await Utils.notify("🌟 New SPOT token found! " + spotMarket.symbol + "\nBuying $1 of it.");
-          const order = {
-            side: "Bid",
-            symbol: spotMarket.symbol,
-            volume: "1",
-          };
-          await OrderController.openOrderSpot(order);
-        }
-      } else {
-        console.log("No new SPOT market found.");
-      }
-
       console.log(
-        "\n⚜️  Strategy evaluated. Possible orders placed and canceled. " +
-          Utils.getFormatedCurrentDateTime(-3) +
-          "\n\n"
+        "\n⚜️  Strategy evaluated. Possible orders placed and canceled. " + Utils.getFormatedCurrentDateTime(-3) + "\n"
       );
     } catch (error) {
       console.log(error);